Move provisioning input analysis to input library
