+++ eng/packages/http-client-csharp-provisioning/generator/Azure.Generator.Provisioning/src/ProvisioningGenerator.cs
@@ -21,6 +21,7 @@ namespace Azure.Generator.Provisioning
     {
         private static ProvisioningGenerator? _instance;
         private ProvisioningOutputLibrary? _outputLibrary;
+        private ProvisioningInputLibrary? _inputLibrary;
 
         /// <summary>
         /// Gets the singleton instance of the provisioning generator.
@@ -41,6 +42,9 @@ namespace Azure.Generator.Provisioning
 
         /// <inheritdoc/>
         public override ProvisioningOutputLibrary OutputLibrary => _outputLibrary ??= new ProvisioningOutputLibrary();
+
+        /// <inheritdoc/>
+        public override ProvisioningInputLibrary InputLibrary => _inputLibrary ??= new ProvisioningInputLibrary(Configuration.OutputDirectory);
 
         /// <inheritdoc/>
         public override ProvisioningTypeFactory TypeFactory { get; }
